requirements.txt
python-telegram-bot==20.7
wikipedia-api==0.6.0